fix: render advertisement element function
