+++ js/main.js
@@ -1,1 +1,5 @@
-import './render-elements.js';
+import { renderAdvertisement } from './render-ad.js';
+
+const mapElement = document.querySelector('#map-canvas');
+
+renderAdvertisement(mapElement, 1);
